+++ lib/utils/HttpUtil.ts
@@ -42,6 +42,9 @@ async function callRetry(func: Function, binding: any, arr: Array<any>, opt: ICa
 function delay(ms: number): Promise<void> {
   return new Promise<void>(a => setTimeout(a, ms))
 }
+async function delayRandom(ms = 3000) {
+  return await delay(Math.ceil(Math.random() * ms))
+}
 
 // 这些错误，不需要重试，暂停，下次可以断点续传
 function isStoppableError(e: Error): boolean {
@@ -63,4 +66,4 @@ function isOssUrlExpired(e: AxiosError): boolean {
   )
 }
 
-export {callRetry, delay, isStoppableError, isNetworkError, isOssUrlExpired}
+export {callRetry, delay, delayRandom, isStoppableError, isNetworkError, isOssUrlExpired}